Include video metadata (so far: duration) in Asset file metadata
This change is made to support a use-case for the Archives Online Channel web app.
We're now embedding videos that have been uploaded to Contentful. We need to display the duration, and prefer to be able to pull that metadata from the DB rather than calculate it on page load.
Therefore, we've added the duration into the file fields.
Other video (or different asset) metadata could also be included in future, if necessary.