refactor: FeatureEnvy in ItemMapper
In generic_fields method a hash is defined and key, values added to it one by one and then returns it.
Now its just returns a hash.

Method assign_field was given a had to which it added new value conditionally.
Now assign_field is renamed as find_entry_for which dont care about the hash but returns the values needed to be added to the original hash.
Its added from the caller method. This fix the FeatureEnvy as the hash is not reffered many time in that method.

Monty's rebase: on rebase with master where I had to merge in changes I extracted out long shared lines to a pprivate method and renamed the method that generates the mapped field by checking relationships.